[Refactor #25] 웨디카이브 수정
-스크랩 취소하기 화면에서 네비게이션 바를 HStack 에서 ToolBar 이용하는 방향으로 수정. (중앙 파트장님이 추천)
-deprecated 수정자들 변경
-스크랩취소하기 화면에서 이미지위에 overlay되어있는 원 버튼 선택 뿐만이 아니라 이미지 선택시에도 선택되도록 변경.
-하위 뷰 파일 분리.
