+++ weady/weady/Sources/Feature/Weadychive/Views/WeadychiveView.swift
@@ -18,7 +18,7 @@ struct WeadychiveView: View {
     @State private var selectedTopTab: TopTab = .curation // 기본 선택 탭
     @State private var showSheet = false // 시트 표시 여부
     @State private var navigateToDeleteView: Bool = false   // 삭제 뷰로 네비게이션 여부
-    // Now managed by ViewModel
+
 
     // MARK: - Body
     var body: some View {
@@ -142,7 +142,7 @@ struct TopTabIndicatorView: View {
                                     .frame(height: 1)
                             }
                         }
-                        .frame(maxWidth: .infinity)  
+                        .frame(maxWidth: .infinity)
                     }
                     .frame(maxWidth: .infinity)
                 }
@@ -243,7 +243,7 @@ struct SheetView: View {
                 }
             } label: {
                 Text("스크랩 취소하기")
-                    .foregroundColor(Color(red: 1, green: 0.23, blue: 0.19)) //피그마에 맞는 시스템 레드로 바꿈
+                    .foregroundStyle(Color(red: 1, green: 0.23, blue: 0.19)) //피그마에 맞는 시스템 레드로 바꿈
                 //그냥 피그마에 있는 속성 그대로 가져옴. (extension에서 못찾음)
                     .frame(maxWidth: .infinity, alignment: .leading)
                     .padding(.horizontal, 20)
@@ -255,169 +255,10 @@ struct SheetView: View {
 }
 
 
-enum DeleteContentType {
-    case curation, weadyboard
-
-    var title: String {
-        switch self {
-        case .curation: return "스크랩한 큐레이션"
-        case .weadyboard: return "스크랩한 웨디보드"
-        }
-    }
-
-    var gridColumns: [GridItem] {
-        switch self {
-        case .curation: return Array(repeating: GridItem(.flexible(), spacing: 2), count: 2)
-        case .weadyboard: return Array(repeating: GridItem(.flexible(), spacing: 2), count: 3)
-        }
-    }
-
-    func imageName(for index: Int) -> String {
-        switch self {
-        case .curation: return index % 2 == 0 ? "curation1" : "curation2"
-        case .weadyboard: return "weadyboard\((index % 7) + 1)"
-        }
-    }
-
-    var imageHeight: CGFloat {
-        switch self {
-        case .curation: return 240
-        case .weadyboard: return 164
-        }
-    }
-}
-
-struct DeleteView: View {
-    @Environment(\.dismiss) var dismiss
-    let type: DeleteContentType
-    @Binding var items: [Int]
-    var onDelete: ([Int]) -> Void
-    @State private var selectedItems: Set<Int> = []
-
-   
-    var body: some View {
-        VStack(spacing: 0) {
-            HStack {
-                Button { dismiss() } label: {
-                    Image(systemName: "chevron.left").foregroundStyle(.black)
-                }
-                Spacer()
-                Text("취소할 항목").font(.headline)
-                Spacer()
-                // 완료 버튼 탭 시: 뷰를 닫고 선택된 항목을 삭제
-                Text("완료")
-                    .fontName(.captionMedium14)
-                    .onTapGesture {
-                        dismiss()
-                        DispatchQueue.main.asyncAfter(deadline: .now() + 0.3) {
-                            onDelete(Array(selectedItems))
-                        }
-                    }
-            }
-            .padding(.horizontal)
-            .padding(.vertical, 16)
-
-            ScrollView {
-                LazyVGrid(columns: type.gridColumns, spacing: 2) {
-                    // 삭제 가능한 항목 리스트 표시
-                    ForEach(items, id: \.self) { index in
-                        ZStack(alignment: .topTrailing) {
-                            Image(type.imageName(for: index))
-                                .resizable()
-                                .aspectRatio(contentMode: .fill)
-                                .frame(height: type.imageHeight)
-                                .clipped()
-                                .overlay(
-                                    selectedItems.contains(index) ? Color.black.opacity(0.3) : Color.clear
-                                )
-                            Button {
-                                if selectedItems.contains(index) {
-                                    selectedItems.remove(index)
-                                } else {
-                                    selectedItems.insert(index)
-                                }
-                            } label: {
-                                Image(systemName: selectedItems.contains(index) ? "checkmark.circle.fill" : "circle")
-                                    .resizable()
-                                    .frame(width: 24, height: 24)
-                                    .foregroundStyle(.white)
-                                    .background(Color.black.opacity(0.6))
-                                    .clipShape(Circle())
-                                    .padding(6)
-                            }
-                        }
-                    }
-                }
-                .padding(.horizontal, 2)
-            }
-        }
-        .navigationBarBackButtonHidden(true)
-    }
-}
 
 
 
 
-// MARK: - NoCurationView (스크랩된 큐레이션 없는 경우)
-struct NoCurationView: View {
-    var body: some View {
-        VStack {
-            VStack(spacing:20) {
-                Text("추천 큐레이션에서 마음에 드는 하루를 담아보세요!")
-                    .fontName(.metaMedium12)
-                    .multilineTextAlignment(.center)
-                    .padding(.horizontal)
 
-                Button(action: {
-                    // TODO: - 큐레이션 탐색 화면으로 이동
-                    //WeadychiveView()
-                }) {
-                    Text("큐레이션 보러가기")
-                        .fontName(.captionSemibold14)
-                        .foregroundStyle(.white)
-                        .padding(.vertical, 10)
-                        .padding(.horizontal, 20)
-                        .background(Color.gray900)
-                        .cornerRadius(8)
-                }
-            }
-            .padding(.top, 141)
-            Spacer()
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-        .padding()
-    }
-}
-
-// MARK: - NoWeadyboardView (스크랩된 웨디보드 없는 경우)
-struct NoWeadyboardView: View {
-    var body: some View {
-        VStack {
-            VStack(spacing: 20) {
-                Text("웨디보드에서는 다른 사람의 하루도 아카이빙할 수 있어요!")
-                    .fontName(.metaMedium12)
-                    .multilineTextAlignment(.center)
-                    .padding(.horizontal)
-
-                Button(action: {
-                    // TODO: - 웨디보드 탐색 화면으로 이동
-                    //WeadyboardView()
-                }) {
-                    Text("웨디보드 보러가기")
-                        .fontName(.captionSemibold14)
-                        .foregroundStyle(.white)
-                        .padding(.vertical, 10)
-                        .padding(.horizontal, 20)
-                        .background(Color.gray900)
-                        .cornerRadius(8)
-                }
-            }
-            .padding(.top, 141)
-            Spacer()
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-        .padding()
-    }
-}
 
 #Preview{WeadychiveView()}
